requirements.txt
paddlepaddle-gpu
# paddlespeech==0.2.0
paddlespeech
# pysoundfile==0.10.3
# yaml==0.2.5
# tqdm==4.64.0
nltk==3.7
visualdl
munch

# absl-py==1.1.0
# albucore==0.0.13
# albumentations==1.4.13
# annotated-types==0.7.0
# anyio==3.6.1
# argon2-cffi==21.3.0
# argon2-cffi-bindings==21.2.0
# asttokens==2.0.5
# attrs==21.4.0
audioread==3.0.1
# Babel==2.10.3
# backcall==0.2.0
# beautifulsoup4==4.11.1
# bleach==5.0.0
# boltons==23.0.0
# brotlipy==0.7.0
# cachetools==5.2.0
# certifi==2024.7.4
# cffi==1.14.6
# chardet==4.0.0
# conda==23.5.2
# conda-package-handling==1.7.3
# cryptography==3.4.7
# cycler==0.11.0
# debugpy==1.6.0
# decorator==5.1.1
# defusedxml==0.7.1
# einops==0.8.0
# entrypoints==0.4
# eval-type-backport==0.2.0
# executing==0.8.3
# fastjsonschema==2.15.3
ffmpeg==1.4
fonttools==4.33.3
# google-auth==2.8.0
# google-auth-oauthlib==0.4.6
# grpcio==1.46.3
# idna==2.10
# imageio==2.34.2
# importlib-metadata==4.11.4
# importlib-resources==5.8.0
# ipykernel==6.15.0
# ipython==8.4.0
# ipython-genutils==0.2.0
# ipywidgets==7.7.0
# jedi==0.18.1
# Jinja2==3.1.2
# joblib==1.4.2
# json5==0.9.8
# jsonpatch==1.33
# jsonpointer==2.1
# jsonschema==4.6.0
# jupyter-client==7.3.4
# jupyter-core==4.10.0
# jupyter-server==1.17.1
# jupyterlab==3.4.3
# jupyterlab-pygments==0.2.2
# jupyterlab-server==2.14.0
# jupyterlab-widgets==1.1.0
# kiwisolver==1.4.3
# kornia==0.7.3
# kornia-rs==0.1.5
# lazy-loader==0.4
librosa==0.9.1
llvmlite==0.41.1
Markdown==3.3.7
# MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
# mistune==0.8.4
# mkl-fft==1.3.1
# mkl-random==1.2.2
# mkl-service==2.4.0
# nbclassic==0.3.7
# nbclient==0.6.4
# nbconvert==6.5.0
# nbformat==5.4.0
# nest-asyncio==1.5.5
# networkx==3.1
# notebook==6.4.12
# notebook-shim==0.1.0
numba==0.58.1
numpy==1.22.4
# oauthlib==3.2.0
# olefile==0.46
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandocfilters==1.5.0
# parso==0.8.3
pexpect==4.8.0
# pickleshare==0.7.5
Pillow==9.1.1
# pip==21.1.3
# platformdirs==4.2.2
pluggy==1.0.0
# pooch==1.8.2
# prometheus-client==0.14.1
# prompt-toolkit==3.0.29
# protobuf==5.27.3
protobuf
# psutil==5.9.1
# ptyprocess==0.7.0
# pure-eval==0.2.2
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycosat==0.6.3
# pycparser==2.20
# pydantic==2.8.2
# pydantic-core==2.20.1
pydub==0.25.1
Pygments==2.12.0
pyOpenSSL==20.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
# PySocks==1.7.1
# python-dateutil==2.8.2
# pytz==2022.1
# PyWavelets==1.4.1
PyYAML==6.0
# pyzmq==23.2.0
requests==2.25.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
# ruamel-yaml-conda==0.15.100
# ruamel.yaml==0.16.12
# ruamel.yaml.clib==0.2.6
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
# Send2Trash==1.8.0
# setuptools==52.0.0.post20210125
six==1.16.0
sniffio==1.2.0
soundfile==0.12.1
soupsieve==2.3.2.post1
stack-data==0.3.0
supervisor==4.2.4
# tensorboard==2.9.1
tensorboard
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorboardX==2.6.2.2
# terminado==0.15.0
# threadpoolctl==3.5.0
# tifffile==2023.7.10
# tinycss2==1.1.1
# tomli==2.0.1
toolz==0.12.0
# torch==1.11.0+cu113
torchattacks
# torchaudio==0.11.0
# torchvision==0.12.0+cu113
# tornado==6.1
# tqdm==4.61.2
tqdm
# traitlets==5.3.0
typing-extensions==4.11.0
urllib3==1.26.6
# wcwidth==0.2.5
# webencodings==0.5.1
# websocket-client==1.3.3
# Werkzeug==2.1.2
wheel==0.36.2
widgetsnbextension==3.6.0
zipp==3.8.0